+++ packages/espresso/android/src/main/java/androidx/test/espresso/flutter/common/Duration.java
@@ -5,8 +5,12 @@
 package androidx.test.espresso.flutter.common;
 
 import static com.google.common.base.Preconditions.checkNotNull;
+import static java.util.concurrent.TimeUnit.MILLISECONDS;
+import static java.util.concurrent.TimeUnit.NANOSECONDS;
+import static java.util.concurrent.TimeUnit.SECONDS;
 
 import java.util.concurrent.TimeUnit;
+import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 /**
@@ -15,10 +19,23 @@ import javax.annotation.Nullable;
  *
  * <p>This class is immutable.
  */
-public final class Duration {
+public final class Duration implements Comparable<Duration> {
 
   private final long quantity;
   private final TimeUnit unit;
+
+  /** Constant for a duration of zero. */
+  public static final Duration ZERO = new Duration(0, NANOSECONDS);
+
+  /** Returns a Duration that represents the given seconds. */
+  public static Duration ofSeconds(long seconds) {
+    return new Duration(seconds, SECONDS);
+  }
+
+  /** Returns a Duration that represents the given milliseconds. */
+  public static Duration ofMillis(long millis) {
+    return new Duration(millis, MILLISECONDS);
+  }
 
   /**
    * Initializes a Duration instance.
@@ -32,11 +49,13 @@ public final class Duration {
   }
 
   /** Returns the amount of time. */
+  @Nullable
   public long getQuantity() {
     return quantity;
   }
 
   /** Returns the time unit. */
+  @Nonnull
   public TimeUnit getUnit() {
     return unit;
   }
@@ -50,6 +69,7 @@ public final class Duration {
    * Returns a new Duration instance that adds this instance to the given {@code duration}. If the
    * given {@code duration} is null, this method simply returns this instance.
    */
+  @Nonnull
   public Duration plus(@Nullable Duration duration) {
     if (duration == null) {
       return this;
@@ -57,5 +77,11 @@ public final class Duration {
     long add = unit.convert(duration.quantity, duration.unit);
     long newQuantity = quantity + add;
     return new Duration(newQuantity, unit);
+  }
+
+  @Override
+  public int compareTo(@Nonnull Duration otherDuration) {
+    long otherQuantity = unit.convert(otherDuration.quantity, otherDuration.unit);
+    return Long.compare(quantity, otherQuantity);
   }
 }