+++ packages/espresso/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Updates minimum supported SDK version to Flutter 3.13/Dart 3.1.
 * Updates compileSdk version to 34.
+* Adds the waitFor action.
 
 ## 0.3.0+7